APD-123 rename direktory
