+++ bnipython/api/oneGatePayment.py
@@ -1,6 +1,6 @@
-from lib.net.httpClient import HttpClient
-from lib.util.utils import generateClientId, generateSignature
-from lib.util.response import responseOGP
+from bnipython.net.httpClient import HttpClient
+from bnipython.util.utils import generateClientId, generateSignature
+from bnipython.util.response import responseOGP
 
 
 class OneGatePayment():
