+++ bnipython/net/httpClient.py
@@ -3,7 +3,7 @@ import json
 import ssl
 import base64
 import requests
-from lib.util.utils import getTimestamp, generateTokenSignature
+from bnipython.util.utils import getTimestamp, generateTokenSignature
 
 
 class HttpClient():
