+++ bnipython/util/utils.py
